Check if reboot required
We run into issues when an OS update/upgrade installs a new kernel but the system
has not yet been rebooted.  This is because DAHDI is built with the linux-headers
for the currently running kernel vs. the headers for the kernel that will be
used after the reboot.  To avoid this conflict we add a check to see if there
is a difference between the running and last/recently installed kernel differ.
If so, we suggest a reboot.