Prevent "apt update; apt upgrade" from updating the kernel
The DAHDI kernel module and the version of the running kernel need to be in sync.
To ensure that an unplanned "apt update; apt upgrade" doesn't break the dependency
we offer/ask the admin if they want to block these updates.  IF they say "yes"
then we "hold" the packages.